fix(sprint-2a): int64 return + timeout for index ctx (Gemini #65)
Addresses Gemini's two MEDIUM findings on PR #65:

1. CountTransactionsNetwork now returns int64 instead of int, matching
   the other count helpers in this package (CountBlocksNetwork,
   CountValidators) and removing the implicit 2^31-1 ceiling on a
   chain whose transactionByAddress collection grows linearly with
   activity. The single caller (routes.go:251) stores the result into
   gin.H which doesn't care about the integer width.

2. initializeCollections now uses context.WithTimeout(60s) instead of
   context.Background(). A stalled MongoDB previously could hang
   syncer startup indefinitely on CreateIndex; bounded so we either
   succeed or fail fast and log loudly.

+++ backendAPI/db/transaction.go
@@ -288,7 +288,9 @@ func ReturnTransactions(address string, page, limit int) ([]models.TransactionBy
 // CountTransactionsNetwork returns the total transactionByAddress row count.
 // Uses EstimatedDocumentCount (metadata read) rather than CountDocuments({}),
 // which is O(rows) on a collection that grows linearly with chain activity.
-func CountTransactionsNetwork() (int, error) {
+// Returns int64 to match sibling count helpers and to avoid an implicit cap
+// at 2^31-1 on long-running chains.
+func CountTransactionsNetwork() (int64, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 
@@ -297,7 +299,7 @@ func CountTransactionsNetwork() (int, error) {
 		return 0, fmt.Errorf("failed to count transactions: %v", err)
 	}
 
-	return int(count), nil
+	return count, nil
 }
 
 func CountTransactions(address string) (int, error) {
